fix: frames are floats, not ints

+++ mktoc/disc.py
@@ -397,7 +397,7 @@ class _TrackTime(object):
       elif isinstance(arg,tuple):
          # assume arg is correct format
          self._time = arg
-      elif isinstance(arg,int):
+      elif isinstance(arg,float):
          # convert frame count to min,sec,frames
          min_,fr = divmod(arg, self._FPM)
          sec,fr = divmod(fr, self._FPS)
